updated both programs to work with docker compose
