+++ .gitignore
@@ -1,1 +1,1 @@
-personal.json
+discord_bot/secrets.json
